+++ Scripts/BTS/Modules/Hive/Services/GetHiveCommand.cs
@@ -6,11 +6,11 @@ using System.Collections.Generic;
 
 namespace BTS {
     internal class GetHiveCommand : BaseNetworkService<GetHiveResponse>, IGetHiveService {
-        private Action<List<UserModel>, int, UserModel, int> m_callback;
+        private Action<List<UserModel>, int, UserModel, float> m_callback;
         public GetHiveCommand() {
         }
 
-        public void Execute(int hiveId, Action<List<UserModel>, int, UserModel, int> callback, int offset = 0, int limit = 0) {
+        public void Execute(int hiveId, Action<List<UserModel>, int, UserModel, float> callback, int offset = 0, int limit = 0) {
             m_callback = callback;
             SendPackage(new BTS_GetHive(hiveId, offset, limit));
         }
