+++ Scripts/BTS/WebServer/Models/PackageResponces/GetHiveResponse.cs
@@ -5,7 +5,7 @@ namespace BTS {
         public List<UserModel> Users { get; private set; }
         public int MembersCount { get; private set; }
         public UserModel Parent { get; private set; }
-        public int TotalImpact { get; private set; }
+        public float TotalImpact { get; private set; }
         public override void Parse(Dictionary<string, object> data) {
             Users = new List<UserModel>();
             var hiveSource = (List<object>)data["hive"];
@@ -20,7 +20,7 @@ namespace BTS {
             }
             MembersCount = int.Parse(data["hive_count"].ToString());
             if (data.ContainsKey("hive_total")) {
-                TotalImpact = int.Parse(data["hive_total"].ToString());
+                TotalImpact = float.Parse(data["hive_total"].ToString());
             }
         }
     }
